pyproject.toml
[tool.poetry]
name = "djangobld"
version = "0.1.0"
description = ""
authors = ["IbrahimE11 <ibrahimemad.2002x@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-model-utils = "^4.2.0"
Pillow = "^9.2.0"
django-imagekit = "^4.1.0"
djangorestframework = "^3.14.0"
Pygments = "^2.13.0"
httpie = "^3.2.1"
django-extensions = "^3.2.1"
django-rest-knox = "^4.2.0"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
